Add files via upload
Had to restart.....again.... this time we will only use one string I was successfull getting the computer to print one charactar from a string with this code. Now we will create a loop that will iterate over each charactar and print that charactar. Then finally, the program will iterate over each charactar, convert the charactar to decimal if it meets the condition and then add that value to a temporary register, and then it will keep doing that until the string iteration is complete. When that is done it will print the value as an output (the output is the value stored in the temporary register).