Add files via upload
We have found how to increment over a string and we have conditions based on whether or not the charactar within the user-inputted string is a space, a character, integer, or the end of the string. Now we will are creating a loop that will check whether or not the charactars (that are not integers) what they are in ascii, and then converting the ascii to decimal and saving it to a temporary register.